Move noSignInRequiredUrls
