+++ src/main/modules/oidc/index.ts
@@ -17,9 +17,8 @@ import {
   SIGN_IN_URL,
   SIGN_OUT_URL,
   SWITCH_TO_SOLE_APPLICATION,
+  noSignInRequiredUrls,
 } from '../../steps/urls';
-
-import { noSignInRequiredUrls } from './noSignInRequiredUrls';
 
 /**
  * Adds the oidc middleware to add oauth authentication
