--- src/main/modules/oidc/noSignInRequiredUrls.ts
@@ -1,15 +1,0 @@
-import {
-  ACCESSIBILITY_STATEMENT_URL,
-  COOKIES_URL,
-  PRIVACY_POLICY_URL,
-  TERMS_AND_CONDITIONS_URL,
-  WEBCHAT_URL,
-} from '../../steps/urls';
-
-export const noSignInRequiredUrls = [
-  ACCESSIBILITY_STATEMENT_URL,
-  WEBCHAT_URL,
-  PRIVACY_POLICY_URL,
-  TERMS_AND_CONDITIONS_URL,
-  COOKIES_URL,
-];
